Remove quote-block transcoder function
The only reason for using this function is to add reveal fragment classes.
However, these are already added as classes to :attr_html when the
parse tree is filtered (`org-re-reveal-filter-parse-tree`, and
`org-re-reveal-append-frag`). Removing the transcoder function enables
other attributes (like cite) to be added with an `#+attr_html:` line,
which will be parsed by `org-html-quote-block`.

Add test for quote block